feat: replace js file to ts
