+++ src/workflow/daily/index.ts
@@ -1,0 +1,8 @@
+/*
+   1日に1回実行されるファイル
+ */
+import dotenv from "dotenv";
+
+// 開発環境での動作確認用に環境変数を読み込む
+dotenv.config();
+
